[Build] extract build iconfont (#1242)
