+++ docs/src/asserts/Logo.sketch
