+++ docs/src/demo-entry.js
@@ -1,4 +1,4 @@
-// This file is auto gererated by build/bin/build-entry.js
+// This file is auto gererated by build/build-entry.js
 import { wrapper } from './demo-common';
 
 export default {
