+++ package.json
@@ -15,15 +15,16 @@
   "scripts": {
     "bootstrap": "yarn || npm i && cd ./packages/vant-css/ && yarn || npm i && cd ../../",
     "dev": "npm run build:file && webpack-serve --config build/webpack.dev.js",
-    "build:file": "node build/bin/build-entry.js",
-    "build:components": "node build/bin/build-components.js --color",
+    "lint": "./node_modules/.bin/eslint ./packages --ext .js,.vue",
+    "build:file": "node build/build-entry.js",
+    "build:components": "node build/build-components.js --color",
     "build:vant-css": "gulp build --gulpfile packages/vant-css/gulpfile.js --color",
     "build:vant": "cross-env NODE_ENV=production webpack --color --config build/webpack.build.js && cross-env NODE_ENV=production webpack -p --color --config build/webpack.build.js",
-    "build:style-entry": "node build/bin/build-style-entry.js",
-    "build:changelog": "sh build/bin/build-changelog.sh",
-    "deploy": "rimraf docs/dist && cross-env NODE_ENV=production webpack --config build/webpack.doc.js && gh-pages -d docs/dist",
-    "dist": "npm run bootstrap && node build/bin/build-lib.js",
-    "lint": "./node_modules/.bin/eslint ./packages --ext .js,.vue",
+    "build:style-entry": "node build/build-style-entry.js",
+    "build:changelog": "sh build/build-changelog.sh",
+    "build:iconfont": "gulp --gulpfile ./build/build-iconfont.js",
+    "build:lib": "npm run bootstrap && node build/build-lib.js",
+    "build:site": "rimraf docs/dist && cross-env NODE_ENV=production webpack --config build/webpack.doc.js && gh-pages -d docs/dist",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "open test/coverage/index.html",
@@ -74,11 +75,15 @@
     "fast-glob": "^2.2.2",
     "fast-vue-md-loader": "^1.0.3",
     "gh-pages": "^1.2.0",
+    "gulp": "^3.9.1",
+    "gulp-iconfont": "^9.2.0",
+    "gulp-iconfont-css": "^2.1.0",
     "html-webpack-plugin": "3.2.0",
     "husky": "^0.14.3",
     "jest": "^23.1.0",
     "jest-serializer-vue": "^2.0.2",
     "lint-staged": "^7.1.3",
+    "md5-file": "^4.0.0",
     "postcss": "^6.0.22",
     "postcss-calc": "^6.0.0",
     "postcss-easy-import": "^3.0.0",
