+++ packages/icon/assert/icons.sketch
