+++ packages/icon/config/index.js
