+++ packages/vant-css/src/icon-local.css
@@ -2,5 +2,5 @@
   font-style: normal;
   font-weight: normal;
   font-family: 'vant-icon';
-  src: url('./vant-icon-243435.ttf') format('truetype');
+  src: url('../icon/vant-icon-9b71af.ttf') format('truetype');
 }
