+++ packages/vant-css/src/icon.css
@@ -4,7 +4,7 @@
   font-style: normal;
   font-weight: normal;
   font-family: 'vant-icon';
-  src: url('https://img.yzcdn.cn/zanui/icon/vant-icon-243435.ttf') format('truetype');
+  src: url('https://img.yzcdn.cn/vant/vant-icon-9b71af.ttf') format('truetype');
 }
 
 .van-icon {
